chore: clear the analyzer warnings the ReSharper sweep left behind
Nine warnings, in two groups.

CS0675 (sign-extended operand in a bitwise-or), 3 sites. BtdFile mixes an
int cell index with a uint quadrant bit, which promotes to long and
sign-extends; the indices are grid offsets (cellY - CellMinY) and never
negative, so an explicit (uint) is value-identical. NpcBoundaryVertexStitcher
packs three 21-bit ints into a 64-bit key but cast only two of them — the
third stayed int and sign-extended through the upper bits, which would
collide cells with a negative Z.

CA1823/S1144 (unused private field), 6 sites. NiTMapItemNextOffset and
NavMeshInfoFormIdOffset are the +0 entries of PDB-derived struct layouts.
Deleting them silences the analyzer but loses the layout, so the offsets
moved into the comment above each block, where the rest of the layout is
already recorded and no longer has a gap at 0.

Full non-incremental build: 0 warnings, 0 errors.

+++ src/BethesdaMultitool/Core/Formats/Esm/Land/Btd/BtdFile.cs
@@ -404,7 +404,9 @@ public sealed class BtdFile : IDisposable
                 {
                     // gcvrMask is intentionally NOT reset here: like the reference, it accumulates the
                     // union of ground-cover presence across the cell's quadrants as they are scanned.
-                    var inner = (y << 1 | (uint)(yc >> (m - 1))) * (NCellsX << 1)
+                    // (uint)y, not y: mixing an int with the uint quadrant bit promotes to long and
+                    // sign-extends. y is a grid index (cellY - CellMinY), so it is never negative.
+                    var inner = ((uint)y << 1 | (uint)(yc >> (m - 1))) * (NCellsX << 1)
                                 + ((x << 1) | (xc >> (m - 1)));
                     var offs = (inner << 3) + _gcvrMapOffs;
                     for (var i = 7; i >= 0; i--, offs++)
@@ -453,7 +455,8 @@ public sealed class BtdFile : IDisposable
         var y = cellY - CellMinY;
         for (var q = 0; q < 4; q++)
         {
-            var offs = ((y << 1 | (uint)(q >> 1)) * (NCellsX << 1) + ((x << 1) | (q & 1))) << 3;
+            // (uint)y as above — the int/uint mix would otherwise sign-extend through the multiply.
+            var offs = (((uint)y << 1 | (uint)(q >> 1)) * (NCellsX << 1) + ((x << 1) | (q & 1))) << 3;
             var t = q << 4;
             for (var i = 0; i < 8; i++)
             {
